Add files via upload
Assignment 2 - BE 602 Data Analysis

Problem 1: What is the nature of association between Brand Commitment and Likelihood to Recommend for the 6 brands included in the data ? 
Problem 2: Amazon is interested in knowing whether the satisfaction with their reward program differs by the gender and household income. 
Problem 3: JC Penny and Amazon are both interested in knowing whether there is a relationship between annual household income and the respective Likelihood to Purchase (time frame) for their brands.  
Problem 4: Does perceptions of Brand Love for JC Penny impact the Likelihood of Recommendation? What is the strength and nature of that relationship?